Update pubgeo-tools recipe
Update pubgeo-tools recipe so that the package includes the following
fixes that are not yet included upstream:

Fix writing offsets text file:
https://github.com/pubgeo/pubgeo/pull/20

Fix error reading TIFF input:
https://github.com/pubgeo/pubgeo/pull/21

The recipe source temporarily points at my fork to include the fix from
PR #20, and the recipe includes a patch for PR #21. Note that a patch
cannot be included for #20 because of
https://github.com/conda/conda-build/issues/1973 ("Cannot patch file
with CRLF line endings").