Loosen gdal/libgdal pinned version
When new patch revisions of the gdal/libgdal packages are released on
conda-forge, the old versions are labelled as 'broken'. These 'broken'
packages aren't found when used as a dependency and don't show up in
searches:

    $ conda search -c conda-forge --override-channels libgdal==2.2.3
    Loading channels: done

    PackagesNotFoundError: The following packages are not available from current channels:

      - libgdal==2.2.3

    Current channels:

      - https://conda.anaconda.org/conda-forge/linux-64
      - https://conda.anaconda.org/conda-forge/noarch

This commit loosens the pinned version of gdal/libgdal to allow patch
updates, as recommend in https://github.com/conda-forge/gdal-feedstock/issues/212.